requirements.txt
asgiref==3.11.0
asttokens==3.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
decorator==5.2.1
Django==6.0
django-admin==2.0.2
django-cors-headers==4.3.1
django-excel-response2==3.0.6
django-extensions==3.2.3
django-filter==25.2
django-redis==5.4.0
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
excel-base==1.0.4
executing==2.2.1
factory-boy==3.3.0
Faker==40.1.0
hiredis==3.3.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
ipython==8.19.0
isoweek==1.3.3
jedi==0.19.2
matplotlib-inline==0.2.1
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.31.0
screen==1.0.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
stripe==7.11.0
TimeConvert==3.0.13
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
wcwidth==0.2.14
xlwt==1.3.0
